Improve test suite
