+++ test/StatementTest.php
@@ -7,45 +7,47 @@ use League\Csv\Reader;
 use League\Csv\RecordSet;
 use League\Csv\Statement;
 use PHPUnit\Framework\TestCase;
-use SplTempFileObject;
 
 /**
  * @group statement
  */
 class StatementTest extends TestCase
 {
-    private $csv;
-
     private $stmt;
-
-    private $expected = [
-        ['john', 'doe', 'john.doe@example.com'],
-        ['jane', 'doe', 'jane.doe@example.com'],
-    ];
 
     public function setUp()
     {
-        $tmp = new SplTempFileObject();
-        foreach ($this->expected as $row) {
-            $tmp->fputcsv($row);
-        }
-
-        $this->csv = Reader::createFromFileObject($tmp);
         $this->stmt = new Statement();
     }
 
     public function testProcess()
     {
-        $records = $this->stmt->process($this->csv);
+        $expected = <<<EOF
+Year,Make,Model,Description,Price
+1997,Ford,E350,"ac, abs, moon",3000.00
+1999,Chevy,"Venture ""Extended Edition""","",4900.00
+1999,Chevy,"Venture ""Extended Edition, Very Large""",,5000.00
+1996,Jeep,Grand Cherokee,"MUST SELL!
+air, moon roof, loaded",4799.00
+EOF;
+        $csv = Reader::createFromString($expected);
+        $records = $this->stmt->process($csv);
         $this->assertInstanceof(RecordSet::class, $records);
-        $this->assertEquals($records, $this->csv->select($this->stmt));
-        $this->assertEquals($records, new RecordSet($this->csv, $this->stmt));
+        $this->assertEquals($records, $csv->select($this->stmt));
+        $this->assertEquals($records, new RecordSet($csv, $this->stmt));
     }
 
     public function testStatementImmutability()
     {
         $this->assertSame($this->stmt->setOffset(0)->setLimit(-1), $this->stmt);
         $this->assertNotSame($this->stmt->setOffset(1), $this->stmt);
+        $this->assertNotSame($this->stmt->setLimit(1), $this->stmt);
+        $this->assertNotSame($this->stmt->addFilter(function ($row) {
+            return count($row) != 4;
+        }), $this->stmt);
+        $this->assertNotSame($this->stmt->addSortBy(function ($rowA, $rowB) {
+            return strcasecmp($rowA[0], $rowB[0]);
+        }), $this->stmt);
     }
 
     public function testConstructorSetAllProperties()
