--- test/StreamFilterTest.php
@@ -1,119 +1,0 @@
-<?php
-
-namespace League\Csv\Test;
-
-use League\Csv\Reader;
-use League\Csv\Writer;
-use lib\FilterReplace;
-use LogicException;
-use PHPUnit\Framework\TestCase;
-use SplFileObject;
-use SplTempFileObject;
-
-/**
- * @group stream
- */
-class StreamFilterTest extends TestCase
-{
-    public function testInitStreamFilterWithWriterStream()
-    {
-        $filter = 'php://filter/write=string.rot13/resource='.__DIR__.'/data/foo.csv';
-        $csv = Reader::createFromPath($filter);
-        foreach ($csv->getIterator() as $row) {
-            $this->assertSame($row, ['john', 'doe', 'john.doe@example.com']);
-        }
-    }
-
-    public function testInitStreamFilterWithReaderStream()
-    {
-        $filter = 'php://filter/read=string.toupper/resource='.__DIR__.'/data/foo.csv';
-        $csv = Reader::createFromPath($filter);
-        foreach ($csv->getIterator() as $row) {
-            $this->assertSame($row, ['JOHN', 'DOE', 'JOHN.DOE@EXAMPLE.COM']);
-        }
-    }
-
-    public function testInitStreamFilterWithBothStream()
-    {
-        $filter = 'php://filter/string.toupper/resource='.__DIR__.'/data/foo.csv';
-        $csv = Reader::createFromPath($filter);
-        foreach ($csv->getIterator() as $row) {
-            $this->assertSame($row, ['JOHN', 'DOE', 'JOHN.DOE@EXAMPLE.COM']);
-        }
-    }
-
-    public function testInitStreamFilterWithSplFileObject()
-    {
-        $this->expectException(LogicException::class);
-        Reader::createFromFileObject(new SplFileObject(__DIR__.'/data/foo.csv'))->getStreamFilterMode();
-    }
-
-    public function testappendStreamFilter()
-    {
-        $csv = Reader::createFromPath(__DIR__.'/data/foo.csv');
-        $csv->appendStreamFilter('string.toupper');
-        foreach ($csv->getIterator() as $row) {
-            $this->assertSame($row, ['JOHN', 'DOE', 'JOHN.DOE@EXAMPLE.COM']);
-        }
-    }
-
-    /**
-     * @expectedException LogicException
-     */
-    public function testFailPrependStreamFilter()
-    {
-        $csv = Reader::createFromFileObject(new SplTempFileObject());
-        $this->assertFalse($csv->isActiveStreamFilter());
-        $this->expectException(LogicException::class);
-        $csv->prependStreamFilter('string.toupper');
-    }
-
-    public function testFailedapppendStreamFilter()
-    {
-        $csv = Writer::createFromFileObject(new SplTempFileObject());
-        $this->assertFalse($csv->isActiveStreamFilter());
-        $this->expectException(LogicException::class);
-        $csv->appendStreamFilter('string.toupper');
-    }
-
-    public function testAddMultipleStreamFilter()
-    {
-        $csv = Reader::createFromPath(__DIR__.'/data/foo.csv');
-        $csv->appendStreamFilter('string.tolower');
-        $csv->prependStreamFilter('string.rot13');
-        $csv->appendStreamFilter('string.toupper');
-        foreach ($csv as $row) {
-            $this->assertSame($row, ['WBUA', 'QBR', 'WBUA.QBR@RKNZCYR.PBZ']);
-        }
-    }
-
-    public function testGetFilterPath()
-    {
-        $csv = Writer::createFromPath(__DIR__.'/data/foo.csv');
-        $csv->appendStreamFilter('string.rot13');
-        $csv->prependStreamFilter('string.toupper');
-        $this->assertFalse($csv->getIterator()->getRealPath());
-    }
-
-    public function testGetFilterPathWithAllStream()
-    {
-        $filter = 'php://filter/string.toupper/resource='.__DIR__.'/data/foo.csv';
-        $csv = Reader::createFromPath($filter);
-        $this->assertFalse($csv->getIterator()->getRealPath());
-    }
-
-    public function testSetStreamFilterWriterNewLine()
-    {
-        stream_filter_register(FilterReplace::FILTER_NAME.'*', '\lib\FilterReplace');
-        $csv = Writer::createFromPath(__DIR__.'/data/newline.csv');
-        $csv->appendStreamFilter(FilterReplace::FILTER_NAME."\r\n:\n");
-        $csv->insertOne([1, 'two', 3, "new\r\nline"]);
-    }
-
-    public function testUrlEncodeFilterParameters()
-    {
-        $csv = Reader::createFromPath(__DIR__.'/data/foo.csv');
-        $csv->appendStreamFilter('convert.iconv.UTF-8/ASCII//TRANSLIT');
-        $this->assertCount(1, $csv->select()->fetchAll());
-    }
-}
